added Layout page.. added components, routes and pages folders.
