+++ src/components/Layout/Layout.js
@@ -1,0 +1,18 @@
+import React from 'react';
+//import Header from '../Header/Header.jsx';
+//import Footer from '../Footer/Footer.jsx';
+import Routers from '../../routes/Routers';
+
+
+const Layout = () => {
+  return (
+    <div>
+        <h1>My new project</h1>
+        <div>
+        <Routers/>
+        </div>
+    </div>    
+  );
+};
+
+export default Layout;
