+++ src/components/NoPage/NoPage.js
@@ -1,5 +1,4 @@
 import React from "react";
-import App from "./App";
 export default function Home() {
   return (
     <div>
